improved handling of abort (ctrl + c) (#349)
* no double mypy in devcontainer

removed ms-python extension and using dmypy which checks the whole workspace (specified targets)

* HaltUser thrown on AsyncMessageAbort

So that scenario stops directly, without waiting for tasks to finish.

* let router do the cleanup instead of worker

* kill task greenlet when aborting a run

better exception logging when something goes wrong in a task.

better "request" statistics for SCENARIO.

* a test project that can be used to test in grizzly

* updates for handling abort (ctrl + c) in async-messaged

* update jinja to patch cve-2024-22195

* fixed linting errors

* fix pyright zmq type warnings in tests

increase timeout for `*no_pymqi*` tests.

* handling more error cases during abort

* exclude unnecessary stuff when doing local install builds with grizzly-cli

* additional handling of error paths when async-messaged is terminated

two new messages sent from worker to master:
- return_code, the workers return code. the highest reported return code will be used by the master when exiting
- sig_trap, message to tell master that a signal has been received and that the test is aborted. this is needed since compose services with healthchecks does not trap signals correctly, which means that `grizzly-cli` wouldn't get correct return code

* let router close the integration when closing down

* improved logging

* correct key for cleaning `client_worker_map`

* if there still is sender and/or receivers, change response action to "DISCONNECTING" so that the worker doesn't stop before all entitites has been disconnected

* fixed unit tests

+++ grizzly/scenarios/__init__.py
@@ -5,18 +5,20 @@ import logging
 from os import environ
 from typing import TYPE_CHECKING, Any, Callable, Optional, Union, cast
 
+from gevent.event import Event
 from locust.exception import LocustError
 from locust.user.sequential_taskset import SequentialTaskSet
 
 from grizzly.context import GrizzlyContext
 from grizzly.exceptions import StopScenario
-from grizzly.gevent import GreenletWithExceptionCatching
+from grizzly.gevent import GreenletFactory
 from grizzly.tasks import GrizzlyTask, grizzlytask
 from grizzly.testdata.communication import TestdataConsumer
 from grizzly.types import ScenarioState
 from grizzly.types.locust import StopUser
 
 if TYPE_CHECKING:  # pragma: no cover
+    from gevent import Greenlet
     from locust.user.task import TaskSet
 
     from grizzly.users import GrizzlyUser
@@ -26,9 +28,9 @@ class GrizzlyScenario(SequentialTaskSet):
     consumer: TestdataConsumer
     logger: logging.Logger
     grizzly: GrizzlyContext
-    task_greenlet: Optional[GreenletWithExceptionCatching]
-    task_greenlet_factory: GreenletWithExceptionCatching
-    abort: bool
+    task_greenlet: Optional[Greenlet]
+    task_greenlet_factory: GreenletFactory
+    abort: Event
     spawning_complete: bool
 
     _task_index: int
@@ -39,8 +41,8 @@ class GrizzlyScenario(SequentialTaskSet):
         self.grizzly = GrizzlyContext()
         self.user.scenario_state = ScenarioState.STOPPED
         self.task_greenlet = None
-        self.task_greenlet_factory = GreenletWithExceptionCatching()
-        self.abort = False
+        self.task_greenlet_factory = GreenletFactory(logger=self.logger, ignore_exceptions=[StopScenario])
+        self.abort = Event()
         self.spawning_complete = False
         self.parent.environment.events.quitting.add_listener(self.on_quitting)
         self._task_index = 0
@@ -115,9 +117,10 @@ class GrizzlyScenario(SequentialTaskSet):
         """When locust is quitting, with abort=True (signal received) we should force the
         running task to stop by throwing an exception in the greenlet where it is running.
         """
-        if self.task_greenlet is not None and kwargs.get('abort', False):
-            self.abort = True
+        if self.task_greenlet is not None and kwargs.get('abort', False) and not self.abort.is_set():
+            self.abort.set()
             self.task_greenlet.kill(StopScenario, block=False)
+            self.logger.debug('killed task (greenlet)')
 
     def get_next_task(self) -> Union[TaskSet, Callable]:
         """Use old way of getting task, so we can reset which task to start from."""
@@ -130,12 +133,13 @@ class GrizzlyScenario(SequentialTaskSet):
 
         return task
 
-    def execute_next_task(self) -> None:
+    def execute_next_task(self, index: int, total: int, description: str) -> None:  # type: ignore[override]
         """Execute task in a greenlet, so that we have the possibility to stop it on demand. Any exceptions
         raised in the greenlet should be caught else where.
         """
         try:
-            self.task_greenlet_factory.spawn_blocking(super().execute_next_task)
+            with self.task_greenlet_factory.spawn_task(super().execute_next_task, index, total, description) as greenlet:
+                self.task_greenlet = greenlet
         finally:
             self.task_greenlet = None
